bert: Add TensorTuple workaround
Grid search wants all our data to be in a list of tuples, but
torch wants all our data to be in tuples of lists (so that tensorification
is more efficient). This data structure supports both access patterns,
by detecting cases when the caller is probbaly trying to access
a tuple of lists and transposing accordingly.